Consolidate FilterRegexFactory into shared DatabaseTools.Common
